src/functions/ingestion/requirements.txt
azure-functions==1.21.3
azure-identity==1.19.0
azure-storage-blob==12.23.1
azure-ai-documentintelligence==1.0.2
azure-search-documents==11.6.0
openai==1.54.5
pyodbc==5.2.0
qdrant-client==1.12.1
requests==2.32.3
